ZOOKEEPER-3441: OWASP is flagging jackson-databind-2.9.9.jar for CVE-…
The JIRA is resolved by updating jackson to version 2.9.9.1.

However in so doing I found that it was impossible to run the
dependency check as the CVE lists would always fail to download,
regardless ant or maven. In researching the issue 5.1.0 of the
dependency checker is now available. That fixes this issue (d/l).

However in so doing a couple new problems arise:

The ant dependency check now fails with a circular dependency issue in
one of the transient libraries
(org.sonatype.ossindex#ossindex-service-client;1.2.0). I was unable to
workaround this issue. As such the ant dependency checker is not able
to update to the new version. I believe we should just stop using it
in favor of the maven one as this seems to be Ivy related, as mvn
works just fine with the same change.

Another problem that arises with the dependency checker version
upgrade is that two new issues are identified:

https://www.cvedetails.com/cve/CVE-2008-7220/
which is resolved with the updated prototype.js

https://www.cvedetails.com/cve/CVE-2019-3826/
which seems like a false positive. Please check my work on this.

After these changes the mvn owasp check passes. The code compiles. I
tested the generated documentation and it seems unaffected by the
prototype.js change, although I could have missed this.

Change-Id: I12c9b3111641b066417fc85b155877af5edf9929

Author: Patrick Hunt <phunt@apache.org>

Reviewers: Enrico Olivelli <eolivelli@apache.org>

Closes #1014 from phunt/ZOOKEEPER-3441

(cherry picked from commit fbda06ba57ae298c3f4244d1bfcec4d07ba865e3 and only need the prototype.js)
ref: ENGESC-10686
Change-Id: I8339330c18b9da707087d0c7063e838268664e5f
(cherry picked from commit 79f1595231a64120de9038191091d12caccb9891)