ZOOKEEPER-3719: Fix C Client compilation issues (SOME PARTS)
This fix is related to CDPD-8500, where we want to make the ZooKeeper
C client code to be compatible with centos8.

We only copied a single fix from ZOOKEEPER-3719, as the other parts
were related to code which is not present in 3.5.5 or was about build
descriptors that are different downstream.

Change-Id: Ie1eb046b6eb230cacaac6daf75694a88f93f37fe